Port PyVRP labelling to the 0.14 Model API
A fresh dependency install resolved pyvrp to 0.14, which replaced the
x/y arguments on add_depot/add_client with explicit Location objects and
swapped Route.visits() for scheduled activities, breaking every labelling
test. Pin the floor to the version whose API we now call.

Two test assertions were also stale: solve_instance defaults to the
paper's NStart rule rather than a single start, and the compact-matrix
size check had no margin at three customers, where the packed header
costs as much as the JSON list it replaces.

+++ src/vrp_diffusion_quantum/inference/policy_support.py
@@ -229,9 +229,7 @@ def _inference_timesteps(
         return list(range(num_timesteps - 1, -1, -1))
     span = num_timesteps - 1
     spaced = [
-        int(round(span * (1.0 - index / (num_inference_steps - 1))))
-        if num_inference_steps > 1
-        else 0
+        round(span * (1.0 - index / (num_inference_steps - 1))) if num_inference_steps > 1 else 0
         for index in range(num_inference_steps)
     ]
     # Rounding can collide on adjacent entries; keep the chain strictly descending.
